Factor all logic out of Meta.tsx into data/actions.ts
`Meta.tsx` was way too big, and it’s bad practice to have logic in view files.

I changed every method in `Meta.tsx` to be a function that takes a state object and returns a `setState()` argument. This means most of the functions are pure and testable, and we can split the file up so it’s easier to find things.